pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-conversation-memory-visualizer"
version = "0.1.0"
description = "Local-first conversation memory analytics engine"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn>=0.30.0",
  "pydantic>=2.8.0",
  "python-multipart>=0.0.9",
  "python-dateutil>=2.9.0",
  "numpy>=1.26.0",
  "scikit-learn>=1.5.0",
  "openai>=1.37.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
